% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRH.R
\name{Append_Markdown_Files}
\alias{Append_Markdown_Files}
\title{Append_Markdown_Files :
  Appends the R-Markdown YML meta data like title,bibliography file etc..
  and generates a R-Markdown file from a normal Markdown file in a new
  output folder.}
\usage{
Append_Markdown_Files(input_file_path, author, title, bib_file)
}
\arguments{
\item{input_file_path}{: String path to the input Markdown File}

\item{author}{: String name(s) of author(s)}

\item{title}{: String title of the document}

\item{bib_file}{: String path to the bib file}
}
\value{
none :
}
\description{
Append_Markdown_Files :
  Appends the R-Markdown YML meta data like title,bibliography file etc..
  and generates a R-Markdown file from a normal Markdown file in a new
  output folder.
}
