% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib-tools.R
\name{bib_handler}
\alias{bib_handler}
\title{applies minimal bibliography over bib entries generated}
\usage{
bib_handler(bib_items)
}
\arguments{
\item{bib_items}{bib entries extracted from extraction function}
}
\value{
bbl_record nested list
}
\description{
which are extracted by the extraction function.
}
