% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor-tools.R
\name{comment_filter}
\alias{comment_filter}
\title{comment filter for latex lines data}
\usage{
comment_filter(data)
}
\arguments{
\item{data}{block of data}
}
\value{
filtered data
}
\description{
removes commented latex lines to avoid wrong reference data
}
