% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-tools.R
\name{convert_algorithm}
\alias{convert_algorithm}
\title{convert algorithm2e diagrams to pdf,png}
\usage{
convert_algorithm(alg_object, article_dir)
}
\arguments{
\item{alg_object}{the figure_block of algorithm}

\item{article_dir}{path to the directory which contains RJ article}
}
\value{
algorithm image
}
\description{
this function will create algorithm2e images to be included
in the RJ-web-article
}
