% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-tools.R
\name{convert_all_pdf_png}
\alias{convert_all_pdf_png}
\title{convert all pdf images to png}
\usage{
convert_all_pdf_png(article_dir, fig_block)
}
\arguments{
\item{article_dir}{path to the article working directory}

\item{fig_block}{block of image data}
}
\value{
modified fig_block
}
\description{
reads figure object to deduce conversion and convert the PDF images
into PNG
}
