% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{convert_tikz}
\alias{convert_tikz}
\title{convert tikz diagrams to pdf,png}
\usage{
convert_tikz(fig_block, article_dir)
}
\arguments{
\item{fig_block}{the figure_block of algorithm}

\item{article_dir}{path to the directory which contains RJ article}
}
\value{
tikz image
}
\description{
this function will create tikz images to be included
in the RJ-web-article
}
