% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-tools.R
\name{convert_to_html}
\alias{convert_to_html}
\title{conversion function with logging success and failure}
\usage{
convert_to_html(dir)
}
\arguments{
\item{dir}{absolute path to the article_dir}
}
\value{
Log file
}
\description{
When applied to a list of article_dirs it will also log
successful and failed slugs.
}
\examples{
# TODO

}
