% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj-article.R
\name{convert_to_markdown}
\alias{convert_to_markdown}
\title{convert to Rmarkdown}
\usage{
convert_to_markdown(article_dir)
}
\arguments{
\item{article_dir}{}
}
\value{

}
\description{
convert to Rmarkdown
}
