% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{convert_to_native}
\alias{convert_to_native}
\title{convert LaTeX wrapper to native pandoc AST}
\usage{
convert_to_native(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
creates a converted markdown file, as well as a pkg_meta.yaml file
}
\description{
Uses pandoc along with several lua filters
found at inst/extdata/filters in texor package
}
\details{
convert latex(wrapper) file to pandoc AST
}
\note{
pandoc (along with lua interpreter) is already installed with
 R-studio, hence if not using R-studio you will need to install pandoc.
 https://pandoc.org/installing.html
}
\examples{
# Note This is a minimal example to execute this function
article_dir <- system.file("examples/article",
                 package = "texor")
texor::include_style_file(article_dir)
texor::convert_to_native(article_dir)
}
