% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-tools.R
\name{convert_to_png}
\alias{convert_to_png}
\title{convert one single pdf file to png}
\usage{
convert_to_png(file_path)
}
\arguments{
\item{file_path}{path to the pdf file}
}
\value{
png file of the same
}
\description{
function to invoke `pdftools:pdf_convert()`

This function is designed to be used internally and is called
by `texor::pdf_to_png(file_dir)` function for converting individual of pdf
image.

Note : The extensions in LaTeX source code will
automatically be changed during pandoc conversion
by a lua filter (refer : inst/extdata/image_filter.lua)
}
\examples{
article_dir <- system.file("examples/pdf_conversion",
                 package = "texor")
texor::convert_to_png(paste0(article_dir,"/normal.pdf"))
}
