% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{copy_all_pdf_png}
\alias{copy_all_pdf_png}
\title{copy all images to web/}
\usage{
copy_all_pdf_png(article_dir, fig_block)
}
\arguments{
\item{article_dir}{path to the article working directory}

\item{fig_block}{block of image data}
}
\value{
modified fig_block
}
\description{
reads figure object to read paths and copy them to /web folder.
}
