% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-tools.R
\name{count_env}
\alias{count_env}
\title{count latex environments}
\usage{
count_env(article_dir, env_name)
}
\arguments{
\item{article_dir}{path to the directory which contains RJ article}

\item{env_name}{name of the environment}
}
\value{
count of the environment
}
\description{
count common environments like table,figure,verbatim etc..
}
\examples{
article_dir <- system.file("examples/article",
                 package = "texor")
figures <- texor::count_env(article_dir, "figure")
print(paste("figure count : ", figures))
}
