% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{create_article}
\alias{create_article}
\title{Create an R Journal article with a modified template for texor.}
\usage{
create_article(name = "test", edit = TRUE)
}
\arguments{
\item{name}{the name of the tex file, will default to "test"}

\item{edit}{Opens the file for editing in RStudio/R GUI.}
}
\description{
Create an R Journal article with a modified template for texor.
}
\details{
Outputs an LaTeX R Journal paper template set of files in the project directory.
}
