% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib-tools.R
\name{export_embeded_bibliography}
\alias{export_embeded_bibliography}
\title{export embedded bibliography to a bbl file}
\usage{
export_embeded_bibliography(article_dir, file_name)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{name of the tex file}
}
\value{

}
\description{
This function will extract the embedded bibliography and store it in .bbl
file
}
