% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{extract_caption}
\alias{extract_caption}
\title{extract caption}
\usage{
extract_caption(figure_lines)
}
\arguments{
\item{figure_lines}{the block of raw figure data from LaTeX document}
}
\value{
caption element
}
\description{
this function will extract \\caption{} from LaTeX figure
environment.
}
