% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{extract_embeded_tikz_image}
\alias{extract_embeded_tikz_image}
\title{extracts embedded tikz image(s) from a tex file}
\usage{
extract_embeded_tikz_image(article_dir, file_name)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{name of the tex file}
}
\value{

}
\description{
This will isolate the tikz image into a new text file which will be
read by a lua filter
}
