% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{extract_label}
\alias{extract_label}
\title{extract label}
\usage{
extract_label(figure_lines)
}
\arguments{
\item{figure_lines}{the block of raw figure data from LaTeX document}
}
\value{
label element
}
\description{
this function will extract \\label{} from LaTeX figure
environment.
}
