% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{extract_multiple_tikz}
\alias{extract_multiple_tikz}
\title{a sub-function to isolate multiple tikz images}
\usage{
extract_multiple_tikz(src_file_data, fig_start_list, fig_end_list)
}
\arguments{
\item{src_file_data}{path to the directory which contains tex article}

\item{fig_start_list}{start points of figures around a tikz image (list)}

\item{fig_end_list}{end points of figures around a tikz image (list)}
}
\value{
tikz image data in a list of list of strings
}
\description{
a sub-function to isolate multiple tikz images
}
