% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{extract_path}
\alias{extract_path}
\title{extract path}
\usage{
extract_path(figure_line)
}
\arguments{
\item{figure_line}{the block of raw figure data from LaTeX document}
}
\value{
path element
}
\description{
this function will extract  from LaTeX figure
environment.
}
