% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{extract_single_tikz}
\alias{extract_single_tikz}
\title{a sub-function to isolate a single tikz image}
\usage{
extract_single_tikz(src_file_data, fig_start, fig_end)
}
\arguments{
\item{src_file_data}{path to the directory which contains tex article}

\item{fig_start}{start point of figure around tikz image (integer)}

\item{fig_end}{end point of figure around tikz image (integer)}
}
\value{
tikz image data in a list of strings
}
\description{
a sub-function to isolate a single tikz image
}
