% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{extract_tikz_style}
\alias{extract_tikz_style}
\title{a sub-function to isolate a single tikz image}
\usage{
extract_tikz_style(fig_data, article_dir, iterator)
}
\arguments{
\item{fig_data}{figure tikz block}

\item{article_dir}{path to the article working directory}

\item{iterator}{tikz image number}
}
\value{
tikz image data in a list of strings
}
\description{
a sub-function to isolate a single tikz image
}
