% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-tools.R
\name{figure_reader}
\alias{figure_reader}
\title{figure reader}
\usage{
figure_reader(article_dir, file_name)
}
\arguments{
\item{article_dir}{the directory path where the file and its dependencies are
located}

\item{file_name}{name of the LaTeX file}
}
\value{
figure blocks
}
\description{
figure reader
}
