% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-tools.R
\name{filter_code_env}
\alias{filter_code_env}
\title{filter code environment}
\usage{
filter_code_env(raw_lines, target, replacement)
}
\arguments{
\item{raw_lines}{a vector of lines from text file}

\item{target}{target environment name}

\item{replacement}{replacement environment name}
}
\value{
modified raw_lines
}
\description{
use stream editor to find and replace
cide environments
}
