% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{find_image_extension}
\alias{find_image_extension}
\title{find image extension}
\usage{
find_image_extension(article_dir, image_path, is_tikz = FALSE)
}
\arguments{
\item{article_dir}{the root directory where the article is located}

\item{image_path}{the value from extract_path()}

\item{is_tikz}{tikz image status, default is FALSE}
}
\value{
extension element
}
\description{
this function will extract extension from LaTeX figure
files
}
