% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{find_pdf_files}
\alias{find_pdf_files}
\title{find pdf files}
\usage{
find_pdf_files(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{

}
\description{
find pdf files
}
