% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-tools.R
\name{find_pkg_references}
\alias{find_pkg_references}
\title{scans LaTeX document for any package references}
\usage{
find_pkg_references(input_file)
}
\arguments{
\item{input_file}{name of the input file}
}
\value{
pkg_meta.yaml file containing meta data
}
\description{
Usually this was handled by rjtools package, since the generated Rmd lacks
these commands (they are transformed to links during pandoc conversion), we
need to add one more step to pre-process.
}
