% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texor-tools.R
\name{get_bib_file}
\alias{get_bib_file}
\title{finds the bib file in directory which is referenced in the article}
\usage{
get_bib_file(article_dir, file_name)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{name of the tex file}
}
\value{
name of bib file (character)
}
\description{
finds the bib file in directory which is referenced in the article
}
