% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{get_journal_details}
\alias{get_journal_details}
\title{get Journal details}
\usage{
get_journal_details(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
journal details in an object
}
\description{
get Journal details
}
\examples{
article_dir <- "/home/user/documents/2022-1/2020-36/"
texor::get_journal_details(article_dir)
}
