% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{get_md_file_name}
\alias{get_md_file_name}
\title{get markdown file name}
\usage{
get_md_file_name(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
markdown file name
}
\description{
get markdown file name
}
\examples{
article_dir <- system.file("examples/article",
                 package = "texor")
dir.create(your_article_folder <- file.path(tempdir(), "tempdir"))
x <- file.copy(from = article_dir, to = your_article_folder,recursive = TRUE,)
your_article_path <- paste(your_article_folder,"article",sep="/")
texor::get_md_file_name(your_article_path)
unlink(your_article_folder,recursive = TRUE)
}
