% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{get_single_tikz_styleset}
\alias{get_single_tikz_styleset}
\title{an internal method to isolate tikz styleset}
\usage{
get_single_tikz_styleset(src_file_data, fig_start, fig_end)
}
\arguments{
\item{src_file_data}{source file contents}

\item{fig_start}{starting point of figure of interest}

\item{fig_end}{end point of figure of interest}
}
\value{
tikz_styleset lines as a list of strings
}
\description{
an internal method to isolate tikz styleset
}
