% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib-tools.R
\name{handle_bibliography}
\alias{handle_bibliography}
\title{function to solve bibliography problems(legacy)}
\usage{
handle_bibliography(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
bibliography links the bibtex file with latex source code or
generates a minimal bibtex file from embedded bibliography and links that
file to the latex file
}
\description{
if bibliography exists in bibtex format then (filename.bib) bibtex file will
be preferred.
else this function will generate a minimal bibliography
}
\note{
Use rebib package to use the convertor, aggregator. This is a
legacy parser with limited support and features
}
