% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj-article.R
\name{include_style_file}
\alias{include_style_file}
\title{Include a Metafix.sty style file}
\usage{
include_style_file(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{

}
\description{
This style file helps texor and pandoc to retain metadata and ease
the conversion process.
}
\examples{
texor::include_style_file(".")
}
