% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib-tools.R
\name{link_bibliography_line}
\alias{link_bibliography_line}
\title{append the tex file with a line to link bibliography}
\usage{
link_bibliography_line(article_dir, file_name)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{file name of the tex document}
}
\value{
appends the tex file with a line to link bibliography
}
\description{
append the tex file with a line to link bibliography
}
