% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{log_setup}
\alias{log_setup}
\title{texor log setup}
\usage{
log_setup(article_dir, file_name, idx)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{name of the log file}

\item{idx}{index of log level}
}
\value{
null
}
\description{
a wrapper function for logger package to set up log file for
logging
}
\examples{
 wd <-  system.file("examples/article", package = "texor")
 texor::log_setup(wd, "log-file.log", 2)
}
