% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib-tools.R
\name{make_bibtex_file}
\alias{make_bibtex_file}
\title{writes bibtex data in a structured format to the .bib file}
\usage{
make_bibtex_file(bibtex_data, file_name)
}
\arguments{
\item{bibtex_data}{a list of minimal bibtex data}

\item{file_name}{name of the tex file}
}
\value{

}
\description{
writes bibtex data in a structured format to the .bib file
}
