% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{make_png_files}
\alias{make_png_files}
\title{Internal function to invoke `pdftools:pdf_convert()`}
\usage{
make_png_files(input_files)
}
\arguments{
\item{input_files}{list of file names to be converted to png}
}
\value{

}
\description{
This function is designed to be used internally and is called
by `texor::pdf_to_png(file_dir)` function for converting a
filtered list of pdf images.
}
\examples{
pdf_file <- system.file("examples/pdf_conversion/normal.pdf",
                 package = "texor")
texor::make_png_files(c(pdf_file))

}
