% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib-tools.R
\name{minimal_bibliography}
\alias{minimal_bibliography}
\title{Main parser logic for converting inbuilt latex bibliography}
\usage{
minimal_bibliography(single_bib_data)
}
\arguments{
\item{single_bib_data}{}
}
\value{
bib_record with unique_id, author and
        title(also containing other info)
}
\description{
A function meant to be used internally or in a sequence as described
in the documentation
}
