% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-tools.R
\name{modify_table_environment}
\alias{modify_table_environment}
\title{function to modify env and commands in TeX using GNU sed}
\usage{
modify_table_environment(file_name)
}
\arguments{
\item{file_name}{name of the Tex file where modifications are to be made}
}
\value{

}
\description{
changes are made to :
1. table* environment to table environment
2. \\multicolumn to \\multicolumnx
These are due to the pandoc's limitations and ease in conversion.
}
\details{
\\multicolumnx is redefined in Metafix.sty as
\\renewcommand{\\multicolumnx}[3]{\\multicolumn{#1}{c}{#3}}
}
\examples{
texor::modify_table_environment("myfile.tex")
}
