% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRH.R
\name{Modify_YAML_Data}
\alias{Modify_YAML_Data}
\title{Modify_YAML_Data}
\usage{
Modify_YAML_Data(input_file, ...)
}
\arguments{
\item{...}{the YAML data that needs to be appended}

\item{input_file:}{The file whose YAML data has to be modified.}
}
\value{

}
\description{
Modify_YAML_Data
}
