% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_version_check}
\alias{pandoc_version_check}
\title{check texor pandoc compatibility
minimum pandoc version required is 2.17}
\usage{
pandoc_version_check()
}
\value{
TRUE if v >= 2.17, else FALSE
}
\description{
check texor pandoc compatibility
minimum pandoc version required is 2.17
}
\examples{
texor::pandoc_version_check()
}
