% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-tools.R
\name{patch_code_env}
\alias{patch_code_env}
\title{patch all code environments}
\usage{
patch_code_env(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
writes modified file and also backs up the old file before modification
}
\description{
This function calls the `filter_code_env()` function
over the following code environments described in Rjournal.sty

1. example
2. example*
3. Sin
4. Sout
5. Sinput
6. Soutput
7. smallverbatim
}
