% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation-tools.R
\name{patch_equations}
\alias{patch_equations}
\title{patch equations}
\usage{
patch_equations(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
modified equations in LaTeX file
}
\description{
this function patches equations (particularly eqnarray)
}
