% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{pdf_to_png}
\alias{pdf_to_png}
\title{pdf_to_png}
\usage{
pdf_to_png(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{

}
\description{
pdf_to_png
}
