% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-tools.R
\name{pre_conversion_statistics}
\alias{pre_conversion_statistics}
\title{pre conversion statistics}
\usage{
pre_conversion_statistics(article_dir, write_yaml = TRUE)
}
\arguments{
\item{article_dir}{path to the directory which contains RJ article}

\item{write_yaml}{write to a yaml file (default = TRUE)}
}
\value{
conversion stat block
}
\description{
count common environments,inlines for debugging purposes
}
\examples{
article_dir <- system.file("examples/article",
                 package = "texor")
texor::pre_conversion_statistics(article_dir,write_yaml = FALSE)
}
