% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{pre_process_tikz}
\alias{pre_process_tikz}
\title{pre-process tex file for handling tikz images}
\usage{
pre_process_tikz(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
outfile.tex a modified latex document
}
\description{
This function will remove tikz images from raw_latex and replace
it with placeholders, which will be filled by a lua filter later.
}
