% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRH.R
\name{Produce_HTML}
\alias{Produce_HTML}
\title{Produce HTML}
\usage{
Produce_HTML(input_file_path)
}
\arguments{
\item{input_file_path}{; String path for the R-Markdown file}
}
\value{
none
}
\description{
Produce HTML
}
