% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{read_tikz_metadata}
\alias{read_tikz_metadata}
\title{parse tikz metadata like label and caption}
\usage{
read_tikz_metadata(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{

}
\description{
Also export it to a new text file which can be read by a lua filter
}
