% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnw-tools.R
\name{rnw_to_rmd}
\alias{rnw_to_rmd}
\title{Sweave to RMarkdown}
\usage{
rnw_to_rmd(input_file, front_matter_type = "vignettes", clean_up = TRUE)
}
\arguments{
\item{input_file}{input Sweave file path}

\item{front_matter_type}{knit output type for the RMarkdown file, default is "vignettes"}

\item{clean_up}{whether to clean up the intermediate files, default is TRUE}
}
\value{
True if R Markdown file successfully generated in the same folder
}
\description{
automated function for converting a single Sweave file to R Markdown file
}
\note{
Use pandoc version greater than or equal to 3.1
}
\examples{
article_dir <- system.file("examples/sweave_article", package = "texor")
dir.create(your_article_folder <- file.path(tempdir(), "tempdir"))
x <- file.copy(from = article_dir, to = your_article_folder,recursive = TRUE)
your_article_path <- paste(your_article_folder,"sweave_article", "example.Rnw",sep="/")
your_article_path <- xfun::normalize_path(your_article_path)
rnw_to_rmd(your_article_path, front_matter_type = "vignettes", clean_up = TRUE)
unlink(your_article_folder, recursive = TRUE)
}
