% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{texor_log}
\alias{texor_log}
\title{log messages for various categories}
\usage{
texor_log(message, category, idx)
}
\arguments{
\item{message}{message to be sent}

\item{category}{category of the log message}

\item{idx}{index of log level}
}
\value{
null
}
\description{
a wrapper function for logging different types of log entries
}
\examples{
 wd <-  system.file("examples/article", package = "texor")
texor::log_setup(wd, "log-file.log", 2)
texor::texor_log("Hello", "INFO", 2)
cat(readLines(paste(wd,"/log-file.log",sep="")),sep="\n")
}
