% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texor-tools.R
\name{write_external_file}
\alias{write_external_file}
\title{quick function to writelines to a file}
\usage{
write_external_file(file_name, mode, raw_text)
}
\arguments{
\item{file_name}{name of text file to write contents to}

\item{mode}{mode of opening}

\item{raw_text}{the text/ list of lines to be written}
}
\value{
create/append/write a new file
}
\description{
quick function to writelines to a file
}
