% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz-tools.R
\name{article_has_tikz}
\alias{article_has_tikz}
\title{Check if article has tikz images or not}
\usage{
article_has_tikz(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{file_name of the tex file}
}
\value{
TRUE/FALSE(boolean)
}
\description{
This simple utiliy function will check for tikzpicture environment
}
