% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{convert_to_markdown}
\alias{convert_to_markdown}
\title{convert latex(wrapper) file to markdown}
\usage{
convert_to_markdown(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\description{
Uses pandoc along with several lua filters
found at inst/extdata/filters in texor package
}
\note{
pandoc (along with lua interpreter) is already installed with
 R-studio, hence if not using R-studio you will need to install pandoc.
 https://pandoc.org/installing.html
}
