% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRH.R
\name{Convert_To_Markdown}
\alias{Convert_To_Markdown}
\title{Convert_To_Markdown :
 This Function converts the LaTeX file into a Markdown file using pandoc
 with the same file name in the same directory}
\usage{
Convert_To_Markdown(input_file_path)
}
\arguments{
\item{input_file_path}{: String Path for the Input File}
}
\value{
none :
}
\description{
Convert_To_Markdown :
 This Function converts the LaTeX file into a Markdown file using pandoc
 with the same file name in the same directory
}
