% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRH.R
\name{Copy_Other_Files}
\alias{Copy_Other_Files}
\title{Copy_Other_Files :
  Copies Supporting documents like images,pdf,bib files into the output
  folder for building the HTML version of the R-Markdown file.}
\usage{
Copy_Other_Files(base_dir_path)
}
\arguments{
\item{base_dir_path}{: String indicating base path for the working directory}
}
\value{
none
}
\description{
Copy_Other_Files :
  Copies Supporting documents like images,pdf,bib files into the output
  folder for building the HTML version of the R-Markdown file.
}
