% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{copy_other_files}
\alias{copy_other_files}
\title{Copy Supporting Documents like images,bib file,etc.}
\usage{
copy_other_files(from_path)
}
\arguments{
\item{from_path}{: String indicating base path for the working directory}
}
\value{
none
}
\description{
Copies supporting documents like images,pdf,bib files into the output
folder for building the HTML version of the R-Markdown file.
}
