% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{find_pdf_files}
\alias{find_pdf_files}
\title{find pdf files in a given directory}
\usage{
find_pdf_files(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains pdf images}
}
\value{
list of filtered pdf files or "NA" string which signals no files.
}
\description{
This function will find all pdf images and also filter out pdf article
}
\examples{
file_dir <- system.file("examples/pdf_conversion/",
             package = "texor")
print(find_pdf_files(file_dir))
}
