% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Modify Markdown to R-markdown}
\usage{
generate_rmd(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
R-markdown file in the web folder
}
\description{
generate rmarkdown file in output folder
}
\examples{
# Note This is a minimal example to execute this function
article_dir <- system.file("examples/article",
                 package = "texor")
texor::include_style_file(article_dir)
texor::convert_to_markdown(article_dir)
texor::generate_rmd(article_dir)
}
