% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Modify Markdown to R-markdown}
\usage{
generate_rmd(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{volume}{volume number (int)}

\item{issue}{issue number (int)}
}
\description{
generate rmarkdown file in output folder
}
\examples{
article_dir <- system.file("examples/article",
                 package = "texor")
texor::generate_rmd(article_dir)
}
