% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{generate_rmd}
\alias{generate_rmd}
\title{generate rmarkdown file in output folder}
\usage{
generate_rmd(markdown_file, volume, issue)
}
\arguments{
\item{markdown_file}{relative path along with name of the markdown file}

\item{volume}{volume number (int)}

\item{issue}{issue number (int)}
}
\description{
generate rmarkdown file in output folder
}
\examples{
article_dir <- system.file("examples/article",
                 package = "texor")
sample_root_path <- paste(article_dir, "2022-1", sep ="/")
dir.create(sample_root_path)
sample_article_path <- paste(sample_root_path, "2022-00", sep ="/")
dir.create(sample_article_path)
texor::generate_rmd(, 1, 1)
}
