% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj-article.R
\name{generate_rmd}
\alias{generate_rmd}
\title{generate rmarkdown file in output folder}
\usage{
generate_rmd(markdown_file, volume, issue)
}
\arguments{
\item{markdown_file}{relative path along with name of the markdown file}

\item{volume}{volume number (int)}

\item{issue}{issue number (int)}
}
\value{

}
\description{
generate rmarkdown file in output folder
}
