% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib-tools.R
\name{handle_bibliography}
\alias{handle_bibliography}
\title{function to solve bibliography problems}
\usage{
handle_bibliography(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{

}
\description{
if bibliography exists in bibtex format then (filename.bib) bibtex file will
be preferred.
else this function will generate a minimal bibliography
}
\examples{
wd <-  system.file("examples/bibliography", package = "texor")
texor::handle_bibliography(wd)
}
