% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{include_style_file}
\alias{include_style_file}
\title{Include a Metafix.sty style file}
\usage{
This function is used to include a custom style file (containing
useful macros for conversion) in the wrapper file
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\description{
Include a Metafix.sty style file
}
\details{
This style file helps texor and pandoc to retain metadata and ease
the conversion process.
}
