% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj-article.R
\name{include_style_file}
\alias{include_style_file}
\title{Include Style File}
\usage{
include_style_file(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{

}
\description{
Include Style File
}
