% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article-tools.R
\name{include_style_file}
\alias{include_style_file}
\title{Include Style file}
\usage{
include_style_file(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
adds Metafix.sty file in the article_dir also includes it in RJwrapper file.
}
\description{
Includes the Metafix.sty style file
}
\details{
This style file helps texor and pandoc to retain metadata and ease
the conversion process.
}
\examples{
article_dir <- system.file("examples/article",
                 package = "texor")
dir.create(your_article_folder <- file.path(tempdir(), "tempdir"))
x <- file.copy(from = article_dir, to = your_article_folder,recursive = TRUE,)
your_article_path <- paste(your_article_folder,"article",sep="/")
texor::include_style_file(your_article_path)
unlink(your_article_folder,recursive = TRUE)
}
