% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-tools.R
\name{latex_to_web}
\alias{latex_to_web}
\title{latex to web}
\usage{
latex_to_web(dir)
}
\arguments{
\item{dir}{directory path}
}
\value{
RJweb article document
}
\description{
automated function for converting a single RJarticle to web
}
\examples{
# TODO

}
