% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{log_setup}
\alias{log_setup}
\title{texor log setup}
\usage{
log_setup(article_dir, file_name, namespace, idx)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{name of the log file}

\item{namespace}{namespace of log file}

\item{idx}{index of log level}
}
\value{
NULL but also creates a log file in the article_dir
}
\description{
a wrapper function for logger package to set up log file for
logging
}
\examples{
dir.create(your_article_folder <- file.path(tempdir(), "exampledir"))
example_files <-  system.file("examples/article", package = "texor")
x <- file.copy(from = example_files,to=your_article_folder,recursive = TRUE)
your_article_path <- paste(your_article_folder,"article",sep="/")
 texor::log_setup(your_article_path, "log-file.log", "texor", 2)
unlink(your_article_folder,recursive = TRUE)
}
