% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{make_png_files}
\alias{make_png_files}
\title{make png file}
\usage{
make_png_files(article_dir, input_files)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{input_files}{list of file names to be converted to png}
}
\value{

}
\description{
make png file
}
