% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_version_check}
\alias{pandoc_version_check}
\title{check texor pandoc compatibility}
\usage{
pandoc_version_check()
}
\value{
TRUE if v >= 3.1, else FALSE
}
\description{
texor package requires minimum pandoc version above or equal to 3.1,
 hence this utility will check for the installation and version status.
}
\examples{
rmarkdown::pandoc_version()

texor::pandoc_version_check()
}
