% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-tools.R
\name{patch_figure_env}
\alias{patch_figure_env}
\title{patch figure environments}
\usage{
patch_figure_env(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
writes modified file and also backs up the old file before modification
}
\description{
This function calls the stream editor to change
figure* to figure
1. figure*
}
