% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-tools.R
\name{pdf_to_png}
\alias{pdf_to_png}
\title{convert pdf images in the directory to png format
using pdftools and poppler utils.}
\usage{
pdf_to_png(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{

}
\description{
Note : The extensions in LaTeX source code will
automatically be changed during pandoc conversion
by a lua filter (refer : inst/extdata/image_filter.lua)
}
\examples{
file_dir <- system.file("extdata/examples/pdf_conversion/",
             package = "texor")
texor::pdf_to_png(file_dir)
}
