% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-tools.R
\name{pre_conversion_statistics}
\alias{pre_conversion_statistics}
\title{pre conversion statistics}
\usage{
pre_conversion_statistics(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains RJ article}
}
\value{
conversion stat block
}
\description{
count common environments,inlines for debugging purposes
}
