% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{produce_html}
\alias{produce_html}
\title{call rmarkdown::render to generate html file}
\usage{
produce_html(input_file_path)
}
\arguments{
\item{input_file_path}{; String path for the R-Markdown file}
}
\value{
none
}
\description{
call rmarkdown::render to generate html file
}
