% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{produce_html}
\alias{produce_html}
\title{call rmarkdown::render to generate html file}
\usage{
produce_html(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
none
}
\description{
call rmarkdown::render to generate html file
}
