% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{texor_log}
\alias{texor_log}
\title{log messages for various categories}
\usage{
texor_log(message, category, idx)
}
\arguments{
\item{message}{message to be sent}

\item{category}{category of the log message}

\item{idx}{index of log level}
}
\value{
NUll, but also appends message to the log file in article_dir
}
\description{
a wrapper function for logging different types of log entries
}
\examples{
dir.create(your_article_folder <- file.path(tempdir(), "exampledir"))
example_files <- system.file("examples/article", package = "texor")
x <- file.copy(from = example_files,to=your_article_folder,recursive = TRUE)
your_article_path <- paste(your_article_folder,"article",sep="/")
texor::log_setup(your_article_path, "log-file.log", "texor" , 2)
texor::texor_log("Hello", "INFO", 2)
cat(readLines(paste(your_article_path,"/log-file.log",sep="")),sep="\n")
unlink(your_article_folder,recursive = TRUE)
}
